refactor: over my page and profile page
